% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums.R
\docType{data}
\name{aitrace-vocabularies}
\alias{aitrace-vocabularies}
\alias{vocabularies}
\alias{RISK_CLASS_LEVELS}
\alias{SYSTEM_LIFECYCLE_STATES}
\alias{EVIDENCE_CATEGORIES}
\alias{EVIDENCE_STATUS_LEVELS}
\alias{RISK_CATEGORIES}
\alias{LIKELIHOOD_LEVELS}
\alias{IMPACT_LEVELS}
\alias{RISK_STATUS_LEVELS}
\alias{MODEL_TYPE_LEVELS}
\alias{DATA_ROLE_LEVELS}
\alias{COMPONENT_KIND_LEVELS}
\alias{INCIDENT_CATEGORIES}
\alias{INCIDENT_SEVERITY_LEVELS}
\alias{INCIDENT_LIFECYCLE_STATES}
\alias{IMPROVEMENT_DOMAINS}
\alias{IMPROVEMENT_TRIGGERS}
\alias{IMPROVEMENT_PRIORITY_LEVELS}
\alias{IMPROVEMENT_LIFECYCLE_STATES}
\alias{EFFECT_DIRECTION_LEVELS}
\alias{VALIDATION_METHOD_LEVELS}
\alias{GOVERNANCE_DECISION_LEVELS}
\alias{DECISION_STATUS_LEVELS}
\alias{CHANGE_TYPE_LEVELS}
\alias{CHANGE_STATUS_LEVELS}
\alias{DEPLOYMENT_STATUS_LEVELS}
\alias{VERIFICATION_RESULT_LEVELS}
\alias{CHANGE_IMPACT_LEVELS}
\alias{OBSERVATION_KIND_LEVELS}
\alias{MONITORING_STATUS_LEVELS}
\alias{DRIFT_RESULT_LEVELS}
\alias{DETECTION_SEVERITY_LEVELS}
\alias{DETECTION_STATUS_LEVELS}
\title{Controlled vocabularies used throughout AITrace}
\format{
An object of class \code{character} of length 4.

An object of class \code{character} of length 14.

An object of class \code{character} of length 14.

An object of class \code{character} of length 6.

An object of class \code{character} of length 12.

An object of class \code{character} of length 5.

An object of class \code{character} of length 5.

An object of class \code{character} of length 5.

An object of class \code{character} of length 9.

An object of class \code{character} of length 6.

An object of class \code{character} of length 6.

An object of class \code{character} of length 9.

An object of class \code{character} of length 4.

An object of class \code{character} of length 9.

An object of class \code{character} of length 13.

An object of class \code{character} of length 8.

An object of class \code{character} of length 4.

An object of class \code{character} of length 14.

An object of class \code{character} of length 3.

An object of class \code{character} of length 7.

An object of class \code{character} of length 10.

An object of class \code{character} of length 5.

An object of class \code{character} of length 11.

An object of class \code{character} of length 8.

An object of class \code{character} of length 4.

An object of class \code{character} of length 4.

An object of class \code{character} of length 4.

An object of class \code{character} of length 8.

An object of class \code{character} of length 3.

An object of class \code{character} of length 3.

An object of class \code{character} of length 3.

An object of class \code{character} of length 4.
}
\usage{
RISK_CLASS_LEVELS

SYSTEM_LIFECYCLE_STATES

EVIDENCE_CATEGORIES

EVIDENCE_STATUS_LEVELS

RISK_CATEGORIES

LIKELIHOOD_LEVELS

IMPACT_LEVELS

RISK_STATUS_LEVELS

MODEL_TYPE_LEVELS

DATA_ROLE_LEVELS

COMPONENT_KIND_LEVELS

INCIDENT_CATEGORIES

INCIDENT_SEVERITY_LEVELS

INCIDENT_LIFECYCLE_STATES

IMPROVEMENT_DOMAINS

IMPROVEMENT_TRIGGERS

IMPROVEMENT_PRIORITY_LEVELS

IMPROVEMENT_LIFECYCLE_STATES

EFFECT_DIRECTION_LEVELS

VALIDATION_METHOD_LEVELS

GOVERNANCE_DECISION_LEVELS

DECISION_STATUS_LEVELS

CHANGE_TYPE_LEVELS

CHANGE_STATUS_LEVELS

DEPLOYMENT_STATUS_LEVELS

VERIFICATION_RESULT_LEVELS

CHANGE_IMPACT_LEVELS

OBSERVATION_KIND_LEVELS

MONITORING_STATUS_LEVELS

DRIFT_RESULT_LEVELS

DETECTION_SEVERITY_LEVELS

DETECTION_STATUS_LEVELS
}
\description{
Every classification field is validated against one of these closed
character vectors. Invalid values raise an \code{aitrace_error_choice}.
}
\keyword{datasets}
