% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aitrace_object.R
\name{bump_version}
\alias{bump_version}
\title{Bump version and append a change-log entry}
\usage{
bump_version(x, ...)
}
\arguments{
\item{x}{An \code{aitrace_object}.}

\item{...}{Not currently used.}
}
\value{
The updated object, with \code{version} incremented and a change-log entry appended.
}
\description{
Bump version and append a change-log entry
}
