% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aitrace_object.R
\name{change_log}
\alias{change_log}
\title{Retrieve the change log as a tibble}
\usage{
change_log(x)
}
\arguments{
\item{x}{An \code{aitrace_object}.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with columns \code{version}, \code{timestamp}, \code{note} and \code{actor}.
}
\description{
Retrieve the change log as a tibble
}
