% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change.R
\name{change_table}
\alias{change_table}
\title{Tabulate changes on a system}
\usage{
change_table(system)
}
\arguments{
\item{system}{An \code{\link[=ai_system]{ai_system()}}.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} summarising the system's changes.
}
\description{
Tabulate changes on a system
}
