% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\docType{data}
\name{COMPARISON_RESULT_LEVELS}
\alias{COMPARISON_RESULT_LEVELS}
\title{Baseline comparison result levels}
\format{
An object of class \code{character} of length 4.
}
\usage{
COMPARISON_RESULT_LEVELS
}
\description{
Baseline comparison result levels
}
\keyword{datasets}
