% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\docType{data}
\name{CRITERION_OPERATOR_LEVELS}
\alias{CRITERION_OPERATOR_LEVELS}
\title{Criterion operators}
\format{
An object of class \code{character} of length 10.
}
\usage{
CRITERION_OPERATOR_LEVELS
}
\description{
Criterion operators
}
\keyword{datasets}
