% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decision.R
\name{decision_table}
\alias{decision_table}
\title{Tabulate decisions on a system}
\usage{
decision_table(system)
}
\arguments{
\item{system}{An \code{\link[=ai_system]{ai_system()}}.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} summarising the system's decisions.
}
\description{
Tabulate decisions on a system
}
