% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{default_approval_policy}
\alias{default_approval_policy}
\title{Default policy used by new registries}
\usage{
default_approval_policy()
}
\value{
An \code{aitrace_policy} object.
}
\description{
Default policy used by new registries
}
