% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\docType{data}
\name{EVALUATION_RESULT_LEVELS}
\alias{EVALUATION_RESULT_LEVELS}
\title{Evaluation result levels}
\format{
An object of class \code{character} of length 3.
}
\usage{
EVALUATION_RESULT_LEVELS
}
\description{
Evaluation result levels
}
\keyword{datasets}
