% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aitrace_object.R
\name{from_json}
\alias{from_json}
\title{Restore from JSON text}
\usage{
from_json(txt)
}
\arguments{
\item{txt}{Character JSON text to restore from.}
}
\value{
The restored AITrace object.
}
\description{
Restore from JSON text
}
