% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/improvement.R
\name{improvement_portfolio}
\alias{improvement_portfolio}
\title{Portfolio view of improvements}
\usage{
improvement_portfolio(system)
}
\arguments{
\item{system}{An \code{\link[=ai_system]{ai_system()}}.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} summarising the system's improvements.
}
\description{
Portfolio view of improvements
}
