% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incident.R
\name{incident_table}
\alias{incident_table}
\title{Tabulate incidents}
\usage{
incident_table(system)
}
\arguments{
\item{system}{An \code{\link[=ai_system]{ai_system()}}.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} summarising the system's incidents.
}
\description{
Tabulate incidents
}
