% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/improvement.R
\name{mark_deployed}
\alias{mark_deployed}
\title{Mark an improvement as deployed}
\usage{
mark_deployed(improvement, actor = NULL)
}
\arguments{
\item{improvement}{An \code{\link[=ai_improvement]{ai_improvement()}}.}

\item{actor}{Optional character identifying who performed the action; recorded in the change log.}
}
\value{
The updated \code{ai_improvement}.
}
\description{
Mark an improvement as deployed
}
