% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observation.R
\name{observation_table}
\alias{observation_table}
\title{Tabulate observations attached to a system}
\usage{
observation_table(system)
}
\arguments{
\item{system}{An \code{\link[=ai_system]{ai_system()}}.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} summarising the system's observations.
}
\description{
Tabulate observations attached to a system
}
