% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{registry_list}
\alias{registry_list}
\title{List systems in the registry}
\usage{
registry_list(registry)
}
\arguments{
\item{registry}{A registry created by \code{\link[=new_registry]{new_registry()}}.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} summarising every registered system.
}
\description{
List systems in the registry
}
