% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk.R
\name{risk_severity}
\alias{risk_severity}
\title{Map likelihood x impact to a risk class}
\usage{
risk_severity(likelihood, impact)
}
\arguments{
\item{likelihood}{One of \link{LIKELIHOOD_LEVELS}.}

\item{impact}{One of \link{IMPACT_LEVELS}.}
}
\value{
A single character risk class, one of \link{RISK_CLASS_LEVELS}.
}
\description{
Map likelihood x impact to a risk class
}
