% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai_system.R
\name{risk_table}
\alias{risk_table}
\title{Tabulate risks}
\usage{
risk_table(system)
}
\arguments{
\item{system}{An \code{\link[=ai_system]{ai_system()}}.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} summarising the system's risks.
}
\description{
Tabulate risks
}
