% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aitrace_object.R
\name{to_json}
\alias{to_json}
\title{Serialise to JSON}
\usage{
to_json(x, pretty = TRUE, ...)
}
\arguments{
\item{x}{An object with a \code{\link[=to_list]{to_list()}} method (typically an \code{aitrace_object}).}

\item{pretty}{If \code{TRUE} (default), pretty-print the JSON output.}

\item{...}{Not currently used.}
}
\value{
A single character string of JSON (a \code{jsonlite::json} scalar).
}
\description{
Serialise to JSON
}
