% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aitrace_object.R
\name{to_yaml}
\alias{to_yaml}
\title{Serialise to YAML}
\usage{
to_yaml(x, ...)
}
\arguments{
\item{x}{An object with a \code{\link[=to_list]{to_list()}} method (typically an \code{aitrace_object}).}

\item{...}{Not currently used.}
}
\value{
A single character string of YAML.
}
\description{
Serialise to YAML
}
